Finish test refactoring
