+++ tests/test_extra/__init__.py
@@ -1,0 +1,1 @@
+"""Test extra selectors not found in the CSS specifications."""
