+++ tests/test_extra/test_attribute.py
@@ -1,0 +1,63 @@
+"""Test attribute selectors."""
+from __future__ import unicode_literals
+from .. import util
+
+
+class TestAttribute(util.TestCase):
+    """Test attribute selectors."""
+
+    MARKUP = """
+    <div id="div">
+    <p id="0">Some text <span id="1"> in a paragraph</span>.</p>
+    <a id="2" href="http://google.com">Link</a>
+    <span id="3">Direct child</span>
+    <pre id="pre">
+    <span id="4">Child 1</span>
+    <span id="5">Child 2</span>
+    <span id="6">Child 3</span>
+    </pre>
+    </div>
+    """
+
+    def test_attribute_not_equal_no_quotes(self):
+        """Test attribute with value that does not equal specified value (no quotes)."""
+
+        # No quotes
+        self.assert_selector(
+            self.MARKUP,
+            'body [id!=\\35]',
+            ["div", "0", "1", "2", "3", "pre", "4", "6"],
+            flags=util.HTML5
+        )
+
+    def test_attribute_not_equal_quotes(self):
+        """Test attribute with value that does not equal specified value (quotes)."""
+
+        # Quotes
+        self.assert_selector(
+            self.MARKUP,
+            "body [id!='5']",
+            ["div", "0", "1", "2", "3", "pre", "4", "6"],
+            flags=util.HTML5
+        )
+
+    def test_attribute_not_equal_double_quotes(self):
+        """Test attribute with value that does not equal specified value (double quotes)."""
+
+        # Double quotes
+        self.assert_selector(
+            self.MARKUP,
+            'body [id!="5"]',
+            ["div", "0", "1", "2", "3", "pre", "4", "6"],
+            flags=util.HTML5
+        )
+
+
+class TestAttributeQuirks(TestAttribute):
+    """Test attribute selectors with quirks."""
+
+    def setUp(self):
+        """Setup."""
+
+        self.purge()
+        self.quirks = True
