+++ tests/test_level4/test_placeholder_shown.py
@@ -1,0 +1,61 @@
+"""Test placeholder shown selectors."""
+from __future__ import unicode_literals
+from .. import util
+
+
+class TestPlaceholderShown(util.TestCase):
+    """Test placeholder shown selectors."""
+
+    def test_placeholder_shown(self):
+        """Test placeholder shown."""
+
+        markup = """
+        <input id="0" placeholder="This is some text">
+        <textarea id="1" placeholder="This is some text"></textarea>
+
+        <input id="2" placeholder="">
+        <input id="3">
+
+        <input id="4" type="email" placeholder="This is some text">
+        <input id="5" type="number" placeholder="This is some text">
+        <input id="6" type="password" placeholder="This is some text">
+        <input id="7" type="search" placeholder="This is some text">
+        <input id="8" type="tel" placeholder="This is some text">
+        <input id="9" type="text" placeholder="This is some text">
+        <input id="10" type="url" placeholder="This is some text">
+        <input id="11" type="" placeholder="This is some text">
+        <input id="12" type placeholder="This is some text">
+
+        <input id="13" type="button" placeholder="This is some text">
+        <input id="14" type="checkbox" placeholder="This is some text">
+        <input id="15" type="color" placeholder="This is some text">
+        <input id="16" type="date" placeholder="This is some text">
+        <input id="17" type="datetime-local" placeholder="This is some text">
+        <input id="18" type="file" placeholder="This is some text">
+        <input id="19" type="hidden" placeholder="This is some text">
+        <input id="20" type="image" placeholder="This is some text">
+        <input id="21" type="month" placeholder="This is some text">
+        <input id="22" type="radio" placeholder="This is some text">
+        <input id="23" type="range" placeholder="This is some text">
+        <input id="24" type="reset" placeholder="This is some text">
+        <input id="25" type="submit" placeholder="This is some text">
+        <input id="26" type="time" placeholder="This is some text">
+        <input id="27" type="week" placeholder="This is some text">
+        """
+
+        self.assert_selector(
+            markup,
+            ":placeholder-shown",
+            ['0', '1', '4', '5', '6', '7', '8', '9', '10', '11', '12'],
+            flags=util.HTML
+        )
+
+
+class TestPlaceholderShownQuirks(TestPlaceholderShown):
+    """Test placeholder shown selectors with quirks."""
+
+    def setUp(self):
+        """Setup."""
+
+        self.purge()
+        self.quirks = True
